network_monitoring_backend/requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary  # PostgreSQL adapter for Python
django-timescale  # For integrating Django with TimescaleDB
pysnmp  # SNMP library